requirements.txt
google-genai
faiss-cpu
yfinance
feedparser
pdfplumber
requests
python-dotenv
numpy
boto3
pydantic
fastapi
uvicorn
